Crate flux_config

Modules

flags

Structs

IncludePattern

This specifies which DefId should be checked. It can be specified via multiple patterns of the form -Finclude=<pattern> and the DefId is checked if it matches any of the patterns. Patterns are checked relative to the current working directory.

InferOpts

Options that change the behavior of refinement type inference locally

PartialInferOpts

Pos

Enums

LeanMode

OverflowMode

PointerWidth

RawDerefMode

SmtSolver

Value

Representation of a TOML value.

Statics

CONFIG_FILE

Functions

allow_raw_deref 🔒

allow_uninterpreted_cast

annots

cache_path

catch_bugs

check_overflow 🔒

config_path 🔒

dump_checker_trace

dump_checker_trace_info

dump_constraint

dump_fhir

dump_rty

full_compilation

ignore_default

include_pattern

is_cache_enabled

lean

lean_dir

lean_project

log_dir

no_panic

pointer_width

scrape_quals 🔒

smt_define_fun

solver 🔒

summary

timings

trusted_default

verify